Changed XUSB interface to work with context abstraction.
There are various advantages to this... the main one
being dealing with controllers that don't have an index.

Some of the implementation changes include removing the limit
of four controllers. The potential XInput API will still
be limited to four controllers. LEDs are only set if the
controller is part of the XInput group of controllers.

Also, and perhaps even more importantly, this change of design
appears to fix the synchronization issue. However, it may still
need some tinkering. There has been zero testing with this commit
and needs to be at least turned on before anything.